make it so that install.sh can be run remotely and locally

+++ install.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 
-echo "Downloading leaf..."
-curl https://raw.github.com/maxmackie/leaf/master/leaf > leaf
-chmod +x ./leaf
-mv ./leaf /usr/local/bin/
-rm -f ./install.sh
-echo "Install complete."
+if [ ! -f ./leaf ];
+then
+    echo "Downloading leaf..."
+    curl https://raw.github.com/maxmackie/leaf/master/leaf > leaf
+    echo "Installing..."
+    chmod +x ./leaf
+    mv ./leaf /usr/local/bin/
+    rm -f ./install.sh
+else
+    echo "Installing..."
+    chmod +x ./leaf
+    cp ./leaf /usr/local/bin/
+fi
+
+echo "Install complete. Get coding bucko."
